Do not show delete button when already cashed up

+++ src/routes/manager/cashUp/[Id]/[cashUpUserId]/CashUpItemManual.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import { page } from "$app/stores";
 	import { createCashUpUserDelete, type DTOCashUpUserItemDTO } from "$lib/api";
 	import { Button } from "@kayord/ui";
 	import { Trash2Icon } from "lucide-svelte";
@@ -12,6 +13,8 @@
 
 	let { item, refetch }: Props = $props();
 
+	const isCashedUp = $derived(Number($page.params.cashUpUserId ?? 0) > 0);
+
 	const deleteItem = async () => {
 		await $mutation.mutateAsync({ id: item.id });
 		refetch();
@@ -21,9 +24,11 @@
 <div class="flex items-center justify-between py-1 w-full">
 	<div class="flex items-center">
 		<span class="text-muted-foreground">{item.cashUpUserItemType?.itemType}</span>
-		<Button class="ml-2 size-6" variant="destructive" size="icon" on:click={deleteItem}>
-			<Trash2Icon class="size-4 text-destructive-foreground" />
-		</Button>
+		{#if !isCashedUp}
+			<Button class="ml-2 size-6" variant="destructive" size="icon" on:click={deleteItem}>
+				<Trash2Icon class="size-4 text-destructive-foreground" />
+			</Button>
+		{/if}
 	</div>
 	<span>R {item.value.toFixed(2)}</span>
 </div>
